+++ src-tauri/src/app_state.rs
@@ -29,6 +29,7 @@ pub struct SharedState {
     pub is_quitting: AtomicBool,
     pub http_client: Client,
     pub available_update: Mutex<Option<crate::portable_updater::UpdateManifest>>,
+    pub update_install_gate: tauri::async_runtime::Mutex<()>,
 }
 
 impl SharedState {
@@ -63,6 +64,7 @@ impl SharedState {
             is_quitting: AtomicBool::new(false),
             http_client,
             available_update: Mutex::new(None),
+            update_install_gate: tauri::async_runtime::Mutex::new(()),
         })
     }
 }